Restore the prebuilt project plugin.
